Refactor-ClustersTable
