+++ frontend/src/components/its/ClustersTable/components/index.ts
@@ -1,0 +1,8 @@
+export { default as TableHeader } from './TableHeader';
+export { default as TableContent } from './TableContent';
+export { default as TablePagination } from './TablePagination';
+export { default as FilterChips } from './FilterChips';
+export { default as ActionMenu } from './ActionMenu';
+export { default as StatusBadge } from './StatusBadge';
+export { default as LabelChip } from './LabelChip';
+export { default as EmptyState } from './EmptyState';
