+++ frontend/src/components/its/ClustersTable/dialogs/index.ts
@@ -1,0 +1,2 @@
+export { default as LabelEditDialog } from './LabelEditDialog';
+export { default as DetachClusterDialog } from './DetachClusterDialog';
